requirements/base.txt
pytz==2021.1
Pillow==8.4.0
Django==4.0
django-environ==0.4.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
whitenoise==5.3.0
beautifulsoup4==4.10.0
requests==2.26.0
